fixing a bug in update post

+++ app/api/post_routes.py
@@ -21,20 +21,67 @@ logger = logging.getLogger(__name__)
 def get_all_posts():
     try:
         posts_query = Post.query
-        paginated_posts = hf.paginate_query(posts_query, 'posts')
+        paginated_posts = hf.paginate_query(posts_query, 'posts', process_item_callback=lambda post, _: post.to_dict())
 
-        # Add user information to each post
+        owner_ids = set(post['owner_id'] for post in paginated_posts['posts'])
+
+        users = User.query.filter(User.id.in_(owner_ids)).all()
+        users_dict = {user.id: user.to_dict() for user in users}
+
         for post in paginated_posts['posts']:
-            user = User.query.get(post['owner_id'])
-            if user:
-                post['user_info'] = user.to_dict()
-            else:
-                post['user_info'] = None
+            post['user_info'] = users_dict.get(post['owner_id'])
 
         return jsonify(paginated_posts)
     except Exception as e:
         logging.error(f"Error fetching all posts: {e}")
         return jsonify({"error": str(e)}), 500
+
+    # try:
+    #     posts_query = Post.query
+    #     paginated_posts = hf.paginate_query(posts_query, 'posts')
+
+    #     # Add user information to each post
+    #     # for post in paginated_posts['posts']:
+    #     #     user = User.query.get(post['owner_id'])
+    #     #     if user:
+    #     #         post['user_info'] = user.to_dict()
+    #     #     else:
+    #     #         post['user_info'] = None
+
+    #     return jsonify(paginated_posts)
+    # except Exception as e:
+    #     logging.error(f"Error fetching all posts: {e}")
+    #     return jsonify({"error": str(e)}), 500
+
+# ***************************************************************
+# Endpoint to Get All Posts Images
+# ***************************************************************
+# @post_routes.route('/all/images')
+# def get_all_posts_images():
+#     try:
+#         posts_query = Post.query
+#         paginated_posts = hf.paginate_query(posts_query, 'posts')
+
+#         simplified_posts_data = []
+#         for post in paginated_posts['posts']:
+#             images = Image.query.filter(Image.post_id == post['id']).all()
+#             for image in images:
+#                 simplified_posts_data.append({
+#                     'post_id': post['id'],
+#                     'owner_id': post['owner_id'],
+#                     'image_url': image.url
+#                 })
+
+
+#         return jsonify({
+#             'posts': simplified_posts_data,
+#             'total_items': paginated_posts['total_items'],
+#             'total_pages': paginated_posts['total_pages'],
+#             'current_page': paginated_posts['current_page']
+#         })
+#     except Exception as e:
+#         logging.error(f"Error fetching all posts: {e}")
+#         return jsonify({"error": str(e)}), 500
 
 # ***************************************************************
 # Endpoint to Get Posts of Current User
@@ -64,34 +111,61 @@ def get_posts_of_current_user():
 # ***************************************************************
 # Endpoint to Get Posts by User ID
 # ***************************************************************
+# @post_routes.route('/user/<int:user_id>')
+# def get_posts_by_user_id(user_id):
+#     try:
+#         user = User.query.get(user_id)
+#         if not user:
+#             return jsonify({"error": "User not found."}), 404
+
+#         posts_query = Post.query.filter(Post.owner_id == user_id)
+#         paginated_posts = hf.paginate_query(posts_query, 'user_posts')
+
+#         return jsonify({
+#             "user_posts": [post.to_dict() for post in paginated_posts['user_posts']],
+#             "total_items": paginated_posts['total_items'],
+#             "total_pages": paginated_posts['total_pages'],
+#             "current_page": paginated_posts['current_page']
+#         })
+
+#     except Exception as e:
+#         logging.error(f"Error fetching posts for user (ID: {user_id}): {e}")
+#         return jsonify({"error": "An error occurred while fetching the posts."}), 500
+
 @post_routes.route('/user/<int:user_id>')
 def get_posts_by_user_id(user_id):
     try:
+        # Fetch user
         user = User.query.get(user_id)
         if not user:
             return jsonify({"error": "User not found."}), 404
 
+        user_info = user.to_dict()
+
         posts_query = Post.query.filter(Post.owner_id == user_id)
-        paginated_posts = hf.paginate_query(posts_query, 'user_posts')
 
-        user_posts = []
-        for post in paginated_posts['user_posts']:
-            if isinstance(post, Post):
-                post_dict = post.to_dict()
-            else:
-                post_dict = post
+        image_ids = [post.image_id for post in posts_query]
+        images = Image.query.filter(Image.id.in_(image_ids)).all()
+        image_url_map = {image.id: image.url for image in images}
 
-            user_posts.append(post_dict)
+        def process_post(post, _):
+            post_dict = post.to_dict()
+            post_dict['image_url'] = image_url_map.get(post.image_id)
+            return post_dict
 
-        result = {
-            "user_info": user.to_dict(),
-            "user_posts": user_posts,
+        paginated_posts = hf.paginate_query(
+            posts_query,
+            'user_posts',
+            process_item_callback=process_post
+        )
+
+        return jsonify({
+            "user_info": user_info,
+            "user_posts": paginated_posts['user_posts'],
             "total_items": paginated_posts['total_items'],
             "total_pages": paginated_posts['total_pages'],
             "current_page": paginated_posts['current_page']
-        }
-        ic(result)
-        return jsonify(result)
+        })
 
     except Exception as e:
         logging.error(f"Error fetching posts for user (ID: {user_id}): {e}")
@@ -99,43 +173,6 @@ def get_posts_by_user_id(user_id):
 
 
 
-# ***************************************************************
-# Endpoint to Get All Posts of the Logged-In User
-# ***************************************************************
-@post_routes.route('/current-user-posts')
-@login_required
-def get_current_user_posts():
-    try:
-        user_id = current_user.id
-        user = User.query.get(user_id)
-        if not user:
-            return jsonify({"error": "User not found."}), 404
-
-        posts_query = Post.query.filter(Post.owner_id == user_id)
-        paginated_posts = hf.paginate_query(posts_query, 'user_posts')
-
-        user_posts = []
-        for post in paginated_posts['user_posts']:
-            if isinstance(post, Post):
-                post_dict = post.to_dict()
-            else:
-                post_dict = post
-
-            user_posts.append(post_dict)
-
-        result = {
-            "user_info": user.to_dict(),
-            "user_posts": user_posts,
-            "total_items": paginated_posts['total_items'],
-            "total_pages": paginated_posts['total_pages'],
-            "current_page": paginated_posts['current_page']
-        }
-        ic(result)
-        return jsonify(result)
-
-    except Exception as e:
-        logging.error(f"Error fetching posts for user (ID: {user_id}): {e}")
-        return jsonify({"error": "An error occurred while fetching the posts."}), 500
 
 # ***************************************************************
 # Endpoint to Get Details of a Post by Id
@@ -155,12 +192,8 @@ def get_post_detail(id):
             return jsonify({"error": "Post not found."}), 404
 
         user = User.query.get(post.owner_id)
-        if user:
-            post_info = post.to_dict()
-            post_info['user_info'] = user.to_dict()
-        else:
-            post_info = post.to_dict()
-            post_info['user_info'] = None
+        post_info = post.to_dict()
+        post_info['user_info'] = user.to_dict() if user else None
 
         return jsonify(post_info)
     except Exception as e:
@@ -198,33 +231,17 @@ def create_post():
     """
     try:
         data = request.json
-        title = data.get('title')
-        description = data.get('description', '')
-        album_id = data.get('album_id')
-        image_url = data.get('image_url')
-
+        title, description, album_id, image_url = data.get('title'), data.get('description', ''), data.get('album_id'), data.get('image_url')
 
         if not title or not image_url:
             return jsonify({'error': 'Title and Image URL are required'}), 400
 
-        # Create the Post
-        post = Post(
-            owner_id=current_user.id,
-            album_id=album_id if album_id else None,
-            title=title,
-            description=description
-        )
-
-        db.session.add(post)
-        db.session.commit()
-
-        # Create the Image
-        image = Image(
-            post_id=post.id,
-            url=image_url
-        )
-
+        image = Image(url=image_url)
         db.session.add(image)
+        db.session.flush()
+
+        post = Post(owner_id=current_user.id, album_id=album_id, title=title, description=description, image_id=image.id)
+        db.session.add(post)
         db.session.commit()
 
         return jsonify(post.to_dict()), 201
@@ -233,6 +250,55 @@ def create_post():
         db.session.rollback()
         print(f"Error in create_post: {e}")
         return jsonify({'error': 'An unexpected error occurred'}), 500
+
+# # ***************************************************************
+# # Endpoint to Create a Post
+# # ***************************************************************
+# @post_routes.route('', methods=["POST"])
+# @login_required
+# def create_post():
+#     """
+#     Create a new post. Only accessible by authenticated users.
+#     Returns:
+#         Response: A JSON object with the newly created post details or an error message.
+#     """
+#     try:
+#         data = request.json
+#         title = data.get('title')
+#         description = data.get('description', '')
+#         album_id = data.get('album_id')
+#         image_url = data.get('image_url')
+
+
+#         if not title or not image_url:
+#             return jsonify({'error': 'Title and Image URL are required'}), 400
+
+#         # Create the Post
+#         post = Post(
+#             owner_id=current_user.id,
+#             album_id=album_id if album_id else None,
+#             title=title,
+#             description=description
+#         )
+
+#         db.session.add(post)
+#         db.session.commit()
+
+#         # Create the Image
+#         image = Image(
+#             post_id=post.id,
+#             url=image_url
+#         )
+
+#         db.session.add(image)
+#         db.session.commit()
+
+#         return jsonify(post.to_dict()), 201
+
+#     except Exception as e:
+#         db.session.rollback()
+#         print(f"Error in create_post: {e}")
+#         return jsonify({'error': 'An unexpected error occurred'}), 500
 
 # ***************************************************************
 # Endpoint to Edit a Post
@@ -249,23 +315,23 @@ def update_post(id):
     """
     try:
         post_to_update = Post.query.get(id)
-        if not post_to_update:
-            return jsonify({"error": "Post not found."}), 404
-        if post_to_update.owner_id != current_user.id:
-            return jsonify(message="Unauthorized"), 403
+        if not post_to_update or post_to_update.owner_id != current_user.id:
+            return jsonify({"error": "Post not found or unauthorized"}), 404
 
         data = request.get_json()
         for key, value in data.items():
-            setattr(post_to_update, key, value)
+            if key != 'image_url':
+                setattr(post_to_update, key, value)
 
-        new_image_url = data.get('image_url')
-        if new_image_url:
-            image_to_update = Image.query.filter_by(post_id=id).first()
+        if 'image_url' in data:
+            image_to_update = Image.query.get(post_to_update.image_id)
             if image_to_update:
-                image_to_update.url = new_image_url
+                image_to_update.url = data['image_url']
             else:
-                new_image = Image(post_id=id, url=new_image_url)
+                new_image = Image(url=data['image_url'])
                 db.session.add(new_image)
+                db.session.flush()
+                post_to_update.image_id = new_image.id
 
         db.session.commit()
         return jsonify(post_to_update.to_dict())
@@ -290,11 +356,8 @@ def delete_post(id):
     """
     try:
         post = Post.query.get(id)
-
-        if not post:
-            return jsonify(error="Post not found"), 404
-        if current_user.id != post.owner_id:
-            return jsonify(error="Unauthorized to delete this post"), 403
+        if not post or current_user.id != post.owner_id:
+            return jsonify(error="Post not found or unauthorized"), 404
 
         db.session.delete(post)
         db.session.commit()
@@ -303,7 +366,6 @@ def delete_post(id):
         logging.error(f"Error deleting post (ID: {id}): {e}")
         db.session.rollback()
         return jsonify(error=f"Error deleting post: {e}"), 500
-
 # ***************************************************************
 # Endpoint to Get Comments of a Post by Id with Pagination
 # ***************************************************************
